changed type getAttempts long->int

+++ Java/Algoritmi/Esercizi_Laboratorio/Settimana2/src/it/unicam/cs/asdl2021/combinationlock/Burglar.java
@@ -62,7 +62,7 @@ public class Burglar {
      *         combinazione, oppure -1 se la cassaforte non è stata ancora
      *         forzata.
      */
-    public long getAttempts() {
+    public int getAttempts() {
         return attempts;
     }
 }
